refactor: 전체 구조 정리
