+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,5 +1,6 @@
-package lotto.domain.lotto;
+package lotto.domain;
 
+import lotto.domain.Lotto;
 import lotto.domain.primitive.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
