+++ src/test/java/lotto/domain/rating/RatingTest.java
@@ -1,6 +1,6 @@
 package lotto.domain.rating;
 
-import lotto.domain.lotto.Lotto;
+import lotto.domain.Lotto;
 import lotto.domain.primitive.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
